video calling on digital event
I have it set so if the event is passed its start time instead of going to the details screen it goes to the video call screen. The check in button adds the current user uid to the attendees list of the event doc. I made my own method for that in the Events Data file so if you have another way of doing it you can change it.